==> notebooks/07_test_recommender.ipynb <==
import sys
sys.path.append("..")

from src.recommender import HybridRecommender

recommender= HybridRecommender(base_path="..")

sample_user_id = recommender.user_encoder.classes_[0]

recommender.recommend(
    raw_user_id=sample_user_id,
    k=10,
    alpha=0.7
)
